+++ client/presentation/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".ui.container.ContainerActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MoveMoveAndroid">
