+++ client/presentation/src/main/java/com/everyone/movemove_android/ui/container/ContainerActivity.kt
@@ -1,11 +1,11 @@
-package com.everyone.movemove_android.ui.main
+package com.everyone.movemove_android.ui.container
 
 import androidx.compose.runtime.Composable
 import com.everyone.movemove_android.base.BaseActivity
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class MainActivity : BaseActivity() {
+class ContainerActivity : BaseActivity() {
     @Composable
     override fun InitComposeUi() {
         MainScreen()
