+++ client/presentation/src/main/java/com/everyone/movemove_android/ui/container/navigation/Navigator.kt
@@ -1,7 +1,5 @@
-package com.everyone.movemove_android.ui.main.navigation
+package com.everyone.movemove_android.ui.container.navigation
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.navigation.NavController
 import androidx.navigation.NavGraph.Companion.findStartDestination
