+++ client/presentation/src/main/java/com/everyone/movemove_android/ui/screens/my/MyScreen.kt
@@ -1,4 +1,4 @@
-package com.everyone.movemove_android.ui.main.my
+package com.everyone.movemove_android.ui.screens.my
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
